Move `ParameterGenerationError` to Talpid package
